Move the AppConfig authentication logic to the Repo class, this enables it to be turned off for testing
